Move the singletons to a proper place
